Fixade Bonus 1, 2 och 4

+++ C++/Bonus 1-25/Bonus2/main.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-	int tal1, tal2;
+	float tal1, tal2;
 	
 	cout << "Första talet ";
 	cin >> tal1;
